translation_system/mcp_servers/llm_mcp/requirements.txt
# MCP Server Requirements
mcp>=1.0.0

# Web Framework
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.0

# LLM Providers
openai>=1.0.0

# Authentication
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0